% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugin.R
\name{vech}
\alias{vech}
\title{Half-Vectorization}
\usage{
vech(..., lower = TRUE, scale = FALSE)
}
\arguments{
\item{...}{one or more matrices to be half-vectorized.}

\item{lower}{use lower or upper half-vectorization}

\item{scale}{whether the lower/upper triangular elements
are scaled}
}
\value{
a matrix
}
\description{
Extension of the utility function
\link[ROI]{vech} performing a half-vectorization on the
given matrices.
}
